+++ source/renderer/app/themes/daedalus/dark-blue.js
@@ -311,6 +311,12 @@ const DARK_BLUE_THEME_CONFIG = {
       'rgba(250, 251, 252, 0.5)',
   },
   delegationSetupWizard: {
+    '--theme-delegation-steps-activation-steps-indicator-color': '#e9f4fe',
+    '--theme-delegation-steps-activation-description-color':
+      'rgba(233, 244, 254, 0.8)',
+    '--theme-delegation-steps-activation-fees-label-color': '#e9f4fe',
+    '--theme-delegation-steps-activation-fees-amount-color': '#ea4c5b',
+    '--theme-delegation-steps-activation-address-value-color': '#e9f4fe',
     '--theme-delegation-steps-choose-stake-pool-checkmark-icon-color':
       '#536370',
     '--theme-delegation-steps-choose-stake-pool-delegated-pools-label-color':
@@ -352,6 +358,11 @@ const DARK_BLUE_THEME_CONFIG = {
     '--theme-delegation-steps-choose-wallet-error-select-options-color':
       '#e9f4fe',
     '--theme-delegation-steps-choose-wallet-steps-indicator-color': '#e9f4fe',
+    '--theme-delegation-steps-confirmation-description-color':
+      'rgba(233, 244, 254, 0.8)',
+    '--theme-delegation-steps-confirmation-fees-label-color': '#e9f4fe',
+    '--theme-delegation-steps-confirmation-fees-amount-color': '#ea4c5b',
+    '--theme-delegation-steps-confirmation-steps-indicator-color': '#e9f4fe',
     '--theme-delegation-steps-intro-content-text-color':
       'rgba(233, 244, 254, 0.8)',
     '--theme-delegation-steps-intro-divider-border-color': '#334152',
@@ -469,8 +480,10 @@ const DARK_BLUE_THEME_CONFIG = {
   },
   navDropdown: {
     '--theme-nav-dropdown-item-text-color': '#e9f4fe',
-    '--theme-nav-dropdown-item-background-color': '#263345',
-    '--theme-nav-dropdown-item-background-color-hover': '#536370',
+    '--theme-nav-dropdown-item-background-color': '#314259',
+    '--theme-nav-dropdown-item-background-color-hover':
+      'rgba(233, 244, 254, 0.07)',
+    '--theme-nav-dropdown-item-color-hover': '#e9f4fe',
   },
   navItem: {
     '--theme-nav-item-background-color': '#263345',
@@ -635,7 +648,6 @@ const DARK_BLUE_THEME_CONFIG = {
     '--theme-staking-stake-pools-title-color': '#e9f4fe',
   },
   staking: {
-    '--theme-staking-background-color': 'rgba(233, 244, 254, 0.1)',
     '--theme-staking-content-background-color': '#263345',
     '--theme-staking-content-border-color': '#263345',
     '--theme-staking-font-color-accent': '#cecfd1',
@@ -646,6 +658,7 @@ const DARK_BLUE_THEME_CONFIG = {
     '--theme-staking-table-border-color': '#c6cdd6',
     '--theme-staking-link-color': 'rgba(32, 181, 107, 1)',
     '--theme-staking-link-color-light': 'rgba(32, 181, 107, 0.5)',
+    '--theme-staking-progress-bar-background-color': 'rgba(233, 244, 254, 0.1)',
     '--theme-staking-progress-stripe-dark-1-background-color': '#3c4852',
     '--theme-staking-progress-stripe-dark-2-background-color': '#536370',
     '--theme-staking-table-body-highlighted-text-color': '#2ab467',
@@ -654,7 +667,8 @@ const DARK_BLUE_THEME_CONFIG = {
     '--theme-staking-wallet-row-border-color': 'rgba(102, 122, 138, 0.3)',
     '--theme-staking-dropdown-item-text-color-hover': '#e9f4fe',
     '--theme-staking-dropdown-item-background-color': '#314259',
-    '--theme-staking-dropdown-item-background-color-hover': '#3a4e63',
+    '--theme-staking-dropdown-item-background-color-hover':
+      'rgba(233, 244, 254, 0.07)',
     '--theme-staking-delegation-center-gear-icon-fill-color':
       'rgba(233, 244, 254, 0.5)',
     '--theme-staking-delegation-center-gear-icon-fill-color-active': '#e9f4fe',